Separate token retriever into its own package
